x86/boot: Convert early_serial_base to unsigned long
As a preparatory of adding flexible serial I/O accessors, convert
early_serial_base to unsigned long to cover all possible bus addresses
on the system.

No functional change intended.

Link: https://lore.kernel.org/r/20250502123145.4066635-2-andriy.shevchenko@linux.intel.com

+++ arch/x86/boot/compressed/misc.h
@@ -127,10 +127,10 @@ extern unsigned char _pgtable[];
 
 #ifdef CONFIG_EARLY_PRINTK
 /* early_serial_console.c */
-extern int early_serial_base;
+extern unsigned long early_serial_base;
 void console_init(void);
 #else
-static const int early_serial_base;
+static const unsigned long early_serial_base;
 static inline void console_init(void)
 { }
 #endif
